Prepare basecode to thin-provisioning and active-active backup
This should, right now, change nothing on the existing workflow, except
new arguments are allowed to specify some thin-provisioning server and
active-active backup target.

Codebase changed to use kind of list of server and not a single fixed
server, in order to chain read and write requests. Right now, when this
list comes in action, only the first element is used and act like old
behavior, which should not change anything in end-user point of view.