First working thin-provisioning implementation
When creating the driver, a read and write server is pushed to a list of
read and write server, this server is the default one.

If any thin provisioning server is set, this one will be pushed to the
read server list.

As soon as a read is requested, the read is performed following the
list. If something is found, this one is returned, otherwise the list
keep going until end is reached.